refactored all three bot classes out into bot.py
accidentally a line

docstring
